# Conflicts:
#	upload/admin_area/setting_basic.php
#	upload/admin_area/styles/cb_2014/layout/main.html
#	upload/cb_install/sql/configs.sql
#	upload/cb_install/sql/language_ENG.sql
#	upload/cb_install/sql/language_FRA.sql
#	upload/cb_install/sql/languages.sql
#	upload/cb_install/sql/structure.sql
#	upload/cb_install/sql/user_levels.sql
#	upload/myaccount.php
#	upload/styles/cb_28/layout/blocks/manage/account_head.html
#	upload/styles/cb_28/layout/channels.html
#	upload/styles/cb_28/layout/videos.html
#	upload/styles/cb_28/theme/js/custom.min.js
#	upload/videos.php
#	upload/watch_video.php

+++ upload/includes/classes/video.class.php
@@ -5,17 +5,21 @@ class Video
     private static $video;
     private $tablename = '';
     private $tablename_categories = '';
+    private $field_id = '';
     private $fields = [];
     private $fields_categories = [];
     private $display_block = '';
     private $display_var_name = '';
     private $search_limit = 0;
+    private $broadcast_option = [];
+    private $status_list= [];
 
     /**
      * @throws Exception
      */
     public function __construct(){
         $this->tablename = 'video';
+        $this->field_id = 'videoid';
         $this->tablename_categories = 'video_categories';
 
         $this->fields = [
@@ -66,17 +70,23 @@ class Video
             ,'subscription_email'
         ];
 
-        $version = Update::getInstance()->getDBVersion();
-        if ($version['version'] > '5.3.0' || ($version['version'] == '5.3.0' && $version['revision'] >= 1)) {
+        if (Update::IsCurrentDBVersionIsHigherOrEqualTo('5.3.0', '1')) {
             $this->fields[] = 'is_castable';
             $this->fields[] = 'bits_color';
         }
-        if ($version['version'] > '5.5.0' || ($version['version'] == '5.5.0' && $version['revision'] >= 305)) {
+        if (Update::IsCurrentDBVersionIsHigherOrEqualTo('5.5.0', '305')) {
             $this->fields[] = 'age_restriction';
         }
         if (Update::IsCurrentDBVersionIsHigherOrEqualTo('5.5.0', '371')) {
             $this->fields[] = 'default_poster';
             $this->fields[] = 'default_backdrop';
+        }
+        if (Update::IsCurrentDBVersionIsHigherOrEqualTo('5.5.1', '273')) {
+            $this->fields[] = 'fov';
+        }
+        // TODO : Update revision
+        if (Update::IsCurrentDBVersionIsHigherOrEqualTo('5.5.1', '279')) {
+            $this->fields[] = 'convert_percent';
         }
 
         $this->fields_categories = [
@@ -90,9 +100,16 @@ class Video
             ,'isdefault'
         ];
 
-        $this->display_block = LAYOUT . '/blocks/video.html';
+        $this->display_block = LAYOUT . '/blocks/videos/video.html';
         $this->display_var_name = 'video';
         $this->search_limit = (int)config('videos_items_search_page');
+
+        $this->broadcast_option = ['public' => lang('vdo_br_opt1'), 'private' => lang('vdo_br_opt2'), 'unlisted' => lang('vdo_broadcast_unlisted'), 'logged' => lang('logged_users_only')];
+    }
+
+    public function getBroadcastOption(): array
+    {
+        return $this->broadcast_option;
     }
 
     public static function getInstance(): self
@@ -178,6 +195,7 @@ class Video
     public function getOne(array $params = [])
     {
         $params['first_only'] = true;
+        $params['get_detail'] = true;
         return $this->getAll($params);
     }
 
@@ -272,6 +290,7 @@ class Video
     public function getAll(array $params = [])
     {
         $param_videoid = $params['videoid'] ?? false;
+        $param_videoids = $params['videoids'] ?? false;
         $param_videokey = $params['videokey'] ?? false;
         $param_userid = $params['userid'] ?? false;
         $param_file_name = $params['file_name'] ?? false;
@@ -297,10 +316,14 @@ class Video
         $param_join_user_profile = $params['join_user_profile'] ?? false;
         $param_not_join_user_profile = $params['not_join_user_profile'] ?? false;
         $param_join_flag= $params['join_flag'];
+        $param_get_detail = $params['get_detail'] ?? false;
 
         $conditions = [];
         if( $param_videoid ){
             $conditions[] = $this->getTableName() . '.videoid = \''.mysql_clean($param_videoid).'\'';
+        }
+        if( $param_videoids ){
+            $conditions[] = $this->getTableName() . '.videoid IN ( '.mysql_clean(implode(', ', $param_videoids)).')';
         }
         if( $param_videokey ){
             $conditions[] = $this->getTableName() . '.videokey = \''.mysql_clean($param_videokey).'\'';
@@ -394,13 +417,13 @@ class Video
             foreach ($this->fields as $field) {
                 $group[] = $this->tablename . '.' . $field;
             }
-
-            $select[] = 'users.username AS user_username';
-            $group[] = 'users.username';
+                $select[] = 'users.username AS user_username';
+                $group[] = 'users.username';
         }
 
         $join = [];
-        if( $version['version'] > '5.5.0' || ($version['version'] == '5.5.0' && $version['revision'] >= 264) ) {
+
+        if (Update::IsCurrentDBVersionIsHigherOrEqualTo('5.5.0', '264') && ($param_get_detail || $param_search || $param_tags)) {
             if( !$param_count ){
                 $types = Tags::getVideoTypes();
                 foreach ($types as $type) {
@@ -412,7 +435,7 @@ class Video
             $join[] = 'LEFT JOIN ' . cb_sql_table('tags') .' ON video_tags.id_tag = tags.id_tag';
         }
 
-        if ($version['version'] > '5.5.0' || ($version['version'] == '5.5.0' && $version['revision'] >= 331)) {
+        if (Update::IsCurrentDBVersionIsHigherOrEqualTo('5.5.0', '331') && ($param_get_detail || $param_category || $param_search)) {
             $join[] = 'LEFT JOIN ' . cb_sql_table('videos_categories') . ' ON ' . $this->getTableName() . '.videoid = videos_categories.id_video';
             $join[] = 'LEFT JOIN ' . cb_sql_table('categories') . ' ON videos_categories.id_category = categories.category_id';
 
@@ -435,7 +458,7 @@ class Video
             $join[] = 'INNER JOIN ' . cb_sql_table($collection_items_table) . ' ON ' . $collection_items_table . '.collection_id = ' . $param_collection_id . ' AND ' . $this->getTableName() . '.videoid = ' . $collection_items_table . '.object_id';
         }
 
-        if (!$param_not_join_user_profile) {
+        if (!$param_not_join_user_profile && $param_get_detail) {
             $join[] = 'LEFT JOIN ' . cb_sql_table('user_profile') . ' ON user_profile.userid = users.userid';
 
             if( !$param_count && Update::IsCurrentDBVersionIsHigherOrEqualTo('5.5.1', '136') ){
@@ -926,6 +949,62 @@ class Video
             'final_results' => $results
         ];
     }
+
+    /**
+     * @throws Exception
+     */
+    public function set(int $id_video, string $field, $value){
+        if( !in_array($field, $this->fields) ){
+            return;
+        }
+
+        $sql = 'UPDATE ' . tbl($this->tablename) . ' SET ' . $field . ' = ' . $value . ' WHERE ' . $this->field_id . ' = ' . $id_video;
+        Clipbucket_db::getInstance()->execute($sql);
+    }
+
+    public static function get_thumbs_preview($vdetails, string $size = '416x260'): string
+    {
+        if( config('enable_video_thumbs_preview') != 'yes'){
+            return '';
+        }
+
+        $thumbs = get_thumb($vdetails, true, $size, 'auto');
+        if( empty($thumbs) ){
+            return '';
+        }
+
+        $nb_thumbs = config('video_thumbs_preview_count');
+        $total = count($thumbs);
+
+        if ($nb_thumbs >= $total) {
+            $result = $thumbs;
+        } else {
+            $result = [];
+            $step = ($total - 1) / ($nb_thumbs - 1);
+
+            for ($i = 0; $i < $nb_thumbs; $i++) {
+                $index = (int) round($i * $step);
+                $result[] = $thumbs[$index];
+            }
+        }
+
+
+        return 'data-thumbs=\'' . json_encode($result) . '\'';
+    }
+
+    /**
+     * @return array
+     * @throws Exception
+     */
+    public function getStatusList(): array
+    {
+        if (empty($this->status_list)) {
+            $row = Clipbucket_db::getInstance()->_select('SHOW COLUMNS FROM '.tbl('video').' LIKE \'status\'')[0];
+            preg_match("/^enum\((.*)\)$/", $row['Type'], $matches);
+            $this->status_list = str_getcsv($matches[1], ",", "'");
+        }
+        return $this->status_list;
+    }
 }
 
 class CBvideo extends CBCategory
@@ -985,6 +1064,7 @@ class CBvideo extends CBCategory
             register_anchor_function('display_restricted', 'in_video_thumb', Video::class);
         }
         register_anchor_function('display_banner', 'in_video_thumb', self::class);
+        register_anchor_function('display_convert_percent', 'in_video_thumb', self::class);
     }
 
     /**
@@ -1008,6 +1088,26 @@ class CBvideo extends CBCategory
         if( !empty($text) ){
             echo '<div class="thumb_banner '.$class.'">' . $text . '</div>';
         }
+    }
+
+    /**
+     * @throws Exception
+     */
+    public static function display_convert_percent($video = [])
+    {
+        if( strtolower($video['status']) != 'processing' ){
+            return;
+        }
+
+        // TODO : Update revision
+        if( !Update::IsCurrentDBVersionIsHigherOrEqualTo('5.5.1', '279') ){
+            return;
+        }
+
+        echo '<div class="processing" data-id="' . $video['videoid'] . '">
+                    <i class="fa fa-spinner fa-spin"></i>
+                    <span>' . $video['convert_percent'] . '%</span>
+              </div>';
     }
 
     /**
@@ -1089,13 +1189,22 @@ class CBvideo extends CBCategory
         ];
 
         $version = Update::getInstance()->getDBVersion();
-        if ($version['version'] > '5.3.0' || ($version['version'] == '5.3.0' && $version['revision'] >= 1)) {
+        if (Update::IsCurrentDBVersionIsHigherOrEqualTo('5.3.0', '1')) {
             $basic_fields[] = 'is_castable';
             $basic_fields[] = 'bits_color';
         }
 
-        if ($version['version'] > '5.5.0' || ($version['version'] == '5.5.0' && $version['revision'] >= 305)) {
+        if (Update::IsCurrentDBVersionIsHigherOrEqualTo('5.5.0', '305')) {
             $basic_fields[] = 'age_restriction';
+        }
+
+        if (Update::IsCurrentDBVersionIsHigherOrEqualTo('5.5.1', '273')) {
+            $basic_fields[] = 'fov';
+        }
+
+        // TODO : Update revision
+        if (Update::IsCurrentDBVersionIsHigherOrEqualTo('5.5.1', '279')) {
+            $basic_fields[] = 'convert_percent';
         }
 
         return $this->set_basic_fields($basic_fields);
@@ -1378,7 +1487,7 @@ class CBvideo extends CBCategory
             }
 
             if (User::getInstance()->hasAdminAccess()) {
-                if (!empty($array['status'])) {
+                if (!empty($array['status']) && in_array($array['status'], Video::getInstance()->getStatusList())) {
                     $query_field[] = 'status';
                     $query_val[] = $array['status'];
                 }
@@ -1415,6 +1524,11 @@ class CBvideo extends CBCategory
                 if (!empty($array['embed_code'])) {
                     $query_field[] = 'embed_code';
                     $query_val[] = $array['embed_code'];
+                }
+
+                if (!empty($array['video_fov']) && in_array($array['video_fov'], [180,360])) {
+                    $query_field[] = 'fov';
+                    $query_val[] = $array['video_fov'];
                 }
             }
             //changes made
@@ -1688,11 +1802,13 @@ class CBvideo extends CBCategory
             }
         } else if ($vdetails['file_type'] === 'hls') {
             $directory_path = DirPath::get('videos') . $vdetails['file_directory'] . DIRECTORY_SEPARATOR . $vdetails['file_name'] . DIRECTORY_SEPARATOR;
-            $files_hls = array_diff(scandir($directory_path), ['.', '..']);
-            foreach ($files_hls as $file_hls) {
-                unlink($directory_path . DIRECTORY_SEPARATOR . $file_hls);
+            if(is_dir($directory_path)) {
+                $files_hls = array_diff(scandir($directory_path), ['.', '..']);
+                foreach ($files_hls as $file_hls) {
+                    unlink($directory_path . DIRECTORY_SEPARATOR . $file_hls);
+                }
+                rmdir($directory_path);
             }
-            rmdir($directory_path);
         }
         e(lang('vid_files_removed_msg'), 'm');
     }
